+++ changelog.md
@@ -2,7 +2,13 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9260]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9261]: https://www.github.com/atecoder/dn.vi.ivi
+
+## [8.1.9261] - 2025-05-10
+- Device Tsp Firmware namespace
+  - Add Firmware Info.
+  - Add script resource manager base class.
+  - Add methods to the Script Source Manager interface to allow implementation of the base class by inheritance.
 
 ## [8.1.9260] - 2025-05-09
 - Add Script Resource Manager Interface.